Implement getPagesToUpdate

+++ src/main/kotlin/codes/ati/fetchlin/domain/Page.kt
@@ -4,7 +4,7 @@ data class Page(
         val id: String,
         var url: String,
         var name: String,
-        var interval: Int,
+        var interval: Long,
         var maxNumberOfRevisions: Int,
         var domElement: String,
         val revisions: List<Revision>
